fix : 아이템 서비스 카테고리 아이템 파일 병합
