+++ item_service/requirements.txt
@@ -1,12 +1,23 @@
+alembic==1.13.0
+bcrypt==4.1.1
 blinker==1.7.0
 click==8.1.7
 Flask==3.0.0
+Flask-Bcrypt==1.0.1
 Flask-Cors==4.0.0
+Flask-JWT-Extended==4.5.3
+Flask-Migrate==4.0.5
+Flask-SQLAlchemy==3.1.1
 greenlet==3.0.1
 importlib-metadata==7.0.0
 itsdangerous==2.1.2
 Jinja2==3.1.2
+load-dotenv==0.1.0
+Mako==1.3.0
 MarkupSafe==2.1.3
+psycopg2-binary==2.9.9
+PyJWT==2.8.0
+python-dotenv==1.0.0
 SQLAlchemy==2.0.23
 typing_extensions==4.8.0
 Werkzeug==3.0.1
